docs + naming
