+++ maitake/src/util.rs
@@ -1,7 +1,7 @@
 use core::ptr::NonNull;
 
-mod wakeset;
-pub(crate) use self::wakeset::WakeSet;
+mod wake_batch;
+pub(crate) use self::wake_batch::WakeBatch;
 
 macro_rules! fmt_bits {
     ($self: expr, $f: expr, $has_states: ident, $($name: ident),+) => {
